JWT authentication and some routes needs authorization to access

+++ backend/Controllers/blogController.js
@@ -10,7 +10,9 @@ async function getBloglikes(id) {
 exports.create = async (req, res) => {
     try {
         const blog = await Blog.create({
-            ...req.body
+            title: req.body.title,
+            content: req.body.content,
+            author: req.userid
         })
         console.log(blog)
         res.json({message: "blog succesfully created"})
@@ -34,9 +36,11 @@ exports.getOne = async (req, res) => {
 }
 
 exports.update = async (req, res) => { // console.log(req.params.id)
+    const _id = req.params.id
+    let blog = await Blog.findById(_id)
+    //check if current logged in user is the owner of the blog
+    if (req.userid === blog.author) return res.sendStatus(403)
     try {
-        const _id = req.params.id
-        let blog = await Blog.findById(_id)
         blog.title = req.body.title
         blog.content = req.body.content
         blog.save().then(() => {
@@ -52,6 +56,10 @@ exports.update = async (req, res) => { // console.log(req.params.id)
     }
 }
 exports.delete = async (req, res) => {
+    const _id = req.params.id
+    let blog = await Blog.findById(_id)
+    //check if current logged in user is the owner of the blog
+    if (req.userid === blog.author) return res.sendStatus(403)
     try {
         const _id = req.params.id
         const blog = await Blog.findByIdAndDelete(_id)
